+++ lab1/MakeFile
@@ -4,7 +4,7 @@ CFLAGS  = -std=gnu99 -g -O2
 all:		sim
 
 sim:		shell.c shell.h isa.h sim.c
-		$(CC) $(CFLAGS) -o sim shell.c sim.c
+		cc -o sim shell.h shell.c sim.c
 
 clean: 
 		rm -f sim
