+++ lab1/sim.c
@@ -11,6 +11,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include "shell.h"
+// #include "shell.c"
 #include "isa.h"
